requirements.txt
streamlit==1.32.0
opik==1.6.8
python-dotenv==1.0.0
requests==2.31.0
openai==1.68.2
